feat: Implement server-side handlers for credentials, project configuration, and snapshots in Percy Storybook addon
